spot routes

+++ backend/db/migrations/20231130232104-create-spot.js
@@ -25,10 +25,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9,7)
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9,7)
       },
       name: {
         type: Sequelize.STRING
